// Client/src/app/interfaces/habitacionesDetails.ts
export default interface HabitacionesDetails{
  idHabitacion: number;
  idHotel: number;
  nombre: string;
  descripcion: string;
  capacidad: number;
  disponibilidad: number;
  cantidadCamas: number;
  calificacion: number;
  cantidadBanos: number;
  tipoCama: string;
  cantidadHabitaciones: number;
  hayCocina: number;
  precio: number;
  nombreHotel: string;
  url: string;
  idServicio: number;
  nombreServicio: string;
  icon:string;
}